requirements.txt
selenium==3.141.0
pandas==1.0.1
mysql_connector_repackaged==0.3.1
